+++ BussinessObject/Models/Ebook.cs
@@ -15,7 +15,7 @@ public partial class Ebook
 
     public string? Description { get; set; }
 
-    public int? Price { get; set; }
+    public decimal? Price { get; set; }
 
     public int? Status { get; set; }
